move interface to different file
